requirements.txt
alembic==1.16.5
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Faker==37.8.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
referencing==0.36.2
rpds-py==0.27.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
python-dotenv===1.1.1